Split Append into {Optimized,Unoptimized}Append
This enables a feature (closure components) in Ember where certain
curlies can dynamically resolve into component definition objects and
need to be rendered as component invocations.

The general strategy to handle that is to change all curlies into (at
least semantically) the equivalent of:


```
{{#if (isComponentDefinition foo)}}{{component foo}}{{else}}{{foo}}{{/if}}
```

However, doing all that extra work and bookkeeping for literally every
single curly would be very slow. Therefore, we take a tiered approach
to handle this.

First of all, in Ember the syntactic requirement for this feature is
that there needs to be at least one dot in the path lookup, e.g.
`... as |f|}}{{f.input}}...`, `{{foo.bar.baz}}` or `{{this.foo}}`. This
allows us to emit the same append opcode (now called `OptimizedAppend`)
for the majority of the append curlies. (By default, all curlies are
optimized, but the consuming environment can override that in
`refineStatement`.)

Even then, there are still a large amount of path lookups with a dot
that will not ever resolve into a component, and doing the extra work
for all those cases would be less than ideal. For that reason, even we
need to emit `UnoptimizedAppend`, we still assume that the values are
going to be simple values.

If we discover that the assumption is not true, i.e. we see that the
value is indeed a component definition at either initial render time or
at update time, we will then depot the opcodes into the expanded form.

See the code comments for more details on how it works.