[GLIMMER2] Make proxies not VOLATILE
This was done previously to be more conservative, but I suspect it is
not actually needed. As long as things you are proxying are dirtied with
a `set`, this should be fine.

`CURRENT_TAG` means if anything within the boundaries of the object
model has changed, then this object _might_ have changed. `VOLATILE_TAG`
means the object you are talking about falls outside of the boundaries
of the object model, therefore you literally cannot tell.

I don’t _think_ that’s needed for proxy. If this is causing problems in
real apps during alpha, we can revert it.