[BUGFIX beta] Fix ember-metal/cache
In #13829, we introduced the ability to pass a custom key function for
the cache. However, we are incorrectly passing `(key, value)` to `set`
when the function expects `(obj, value)`.

This is causing problems for caches that uses a custom cache keying
function (such as Glimmer's compilation cache), because the get side
would handle the keying correctly, but the set side would compute the
key based on the already-computed key (i.e. `this.key(this.key(obj))`).

In Glimmer's compilation cache, this is causing the cache to always miss
because the `set` side would cache everything under the `"undefined"`
key.

(cherry picked from commit 80f0462b9b2c749ec52cf096022330de3fb536f3)