Expose `Ember.ApplicationInstance.setupRegistry` for tooling.
Glimmer requires that a number of factories and injections added by
`ApplicationInstance#setupRegistry` are present, but the testing harness
only has access to calling `Ember.Application.buildRegistry` which
leaves the registry in a partially populated state.

This exposes `Ember.ApplicationInstance.setupRegistry` so that
`ember-test-helpers` can ensure that its registry is properly populated.