Remove `InternedString` (and get back on TS next)
The intent of this distinction was to differentiate between possible
ropes and strings, and make sure we convert ropes into strings for hot
paths. However, since JavaScript does not provide an official mechanism
for that conversion, and the Known Trick™ for doing that is too slow,
the experiment didn’t really work out. (We commented out the body of
`intern` for a while.)

Recent versions of TypeScript broke the branding trick on strings, so we
decided it’s time to remove this.