[GLIMMER2] (╯°□°)╯︵ ƃuᴉdǝǝʞ-ʞooq ʎxoɹd dᴉlℲ
This commit flips the responsibilities around – instead of making each
reference do the work to track whether they are dealing with proxies, we
can just make proxies themselves provide the right tag and do the book-
keeping internally.

With this change, `tagFor(proxy)` will return a tag that accounts for
changes on the proxy itself as well as its content. This frees the rest
of the system from worrying about that difference and removes a lot of
runtime `isProxy` checks.