Redefine `isConst` to be tag-based
Since the tag cannot change once initialized, having a `CONSTANT_TAG`
(as opposed to an `UpdatableTag` of `CONSTANT_TAG`) is the same as being
a “const reference”.

This expands the `ConstReference` optimizations we do to more places.
For example, `{{component (concat “foo” “-” “bar”)}}` is now the same as
`{{component “foo-bar”}}`, which is the same as `{{foo-bar}}`.