Ensure dist is built when referencing via git SHA.
In order to allow consumers to `require('glimmer-engine')` a `dist/`
build is required. Normally, this is done as part of the publishing
process for tagged builds, but in order to allow consumers to use git
SHA's and still `require('glimmer-engine')` (which is required for the
Ember build to precompile Ember's internal templates) we are adding a
new `postinstall` script that checks if `dist/` is present and if not
run the build.