Initialize TryOpcode when emulating `enterWithKey`
The code here is to emulate the `enterWithKey` path on the append VM,
however, it is missing the corresponding `exit`, i.e. we are not
initializing the tag on the `TryOpcode`, incorrectly leaving it as an
`UpdatableTag(CONSTANT_TAG)`.

When combined with other update-side optimizations like J-I-N-M, this
causes things to not update correctly.

This is tested via https://github.com/emberjs/ember.js/pull/14082 in
Ember. Unfortunately, this is not currently easily testable on the
Glimmer side due to our reliance on `VOLATILE_TAG`s in glimmer-obj-ref.