Fix multiple selected options in a multi-select
According to the HTML5 tree construction spec, the node must be created
and its attributes set before it is inserted into the DOM. It must be
inserted into the DOM before its children are inserted into it.

This matters in practice because if you create an <option> and insert
it into a <select>, its selection state depends on whether the <select>
is a multiple select. So if you make a <select>, insert an <option>, and
only later assign `multiple`, you get the wrong behavior.

This change aligns more closely with the tree construction spec, which
fixes any bugs caused by the misalignment (in addition to this one that
we know about).