Distinguish static and dynamic attributes
Take 2 of #266, which got reverted in #272.

The wire format separates attributes whose values are purely static
(e.g. href="some.html") from those whose values are dynamic (e.g.
href="{{url}}.html").

Over time, this distinction got lost further down the pipeline, meaning
that static attributes were still going through quite a bit of dynamic
machinery.

This commit restores the distinction all the way through to the opcode
and element stack, which reduces the amount of work needed for static
attribute (it's now very close to "set the attribute with the supplied
value", which is all that's needed for cases where the user wrote
normal HTML syntax).

The main difference between this and #266 is that this still goes
through the "ElementOperations" code, instead of directly going to
`.setAttribute` on the DOM element. The previous approach broke cases
in Ember where the curly component compiler added multiple classes
(e.g. static "ember-view" + dynamic classNameBindings), which requires
the merging logic. This regression is tested in a new test here.