Ensure view registry entries can be unregistered properly.
When the `register` / `unregister` functions were moved from the
`ViewSupportMixin` to the renderer, I didn't properly update them to use
the provided `view`'s `elementId` and was instead using `this.elementId`
(on the renderer) which was obviously undefined.

This meant that we were holding on to a reference of all
views/components ever created (since they were never being released from
the `viewRegistry`). :sob: