[BUGFIX beta] Prevent issues with revalidation during teardown.
When the toplevelView (aka `ownerView`) is being destroyed (during test
teardown generally) an error was ocurring under certain "special"
circumstances. Specifically, if removal of a sibling node triggered a
revalidation of one still attached.

Consider the following example:

```hbs
{{#x-select as |select|}}
  {{#select.option value="1"}}{/select.option}}
  {{#select.option value="2"}}{/select.option}}
{{/x-select}}
```

The components in question are using the registration pattern so that
the children notify/register with the parent upon `didInsertElement`
(and `willDestroyElement`).

When a new option is added or removed from the parent, it updates its
`value` property which is bound to each options `selected`
attribute (to toggle the selection state in the DOM).

The specific mechanism that causes the DOM to get updated when the
various computed properties change is that a Stream object calls
`ownerNode.emberView.scheduleRevalidate`.  This tells the rest of the
system to start walking the tree to flush out any updates to the DOM.

When a view is being removed, we set the `emberView`
property of its `renderNode` to `null` before traversing the subtree of
children. This means, that as the children are removed (and therefore
cause the `value` of the parent to change) the `selected` attribute
binding attempts to call `ownerNode.emberView.scheduleRevalidation()`.

Unfortunately, when we are actually tearing down the
`ownerNode.emberView` itself that invocation results in an error (
cannot call `scheduleRevalidation` on `null`).

---

This change adds a simple guard to avoid calling `scheduleRevalidation`
when the `ownerNode.emberView` is being torn down.

(cherry picked from commit 2fa40155c62315c5c5db9c18bcd409cfad288962)