[BUGFIX beta] Ensure injections happen in engine instances.
Prior to this `_environment` could not be relied upon in routes or
views (which is used when `shouldRender: false` is used with `visit`).

The refactor uses the environment to determine what injections to setup
in the engine-instance, and has the `application-instance` defer to the
`engine-instance` to avoid duplication.

(cherry picked from commit 3de2360772f1f542c06f46e62d096b269e77361b)