[GLIMMER] Ensure all components have injections.
Prior to this change we were defaulting to `Ember.Component` which does
not receive injections (which meant it was missing the `renderer`).
This worked "ok" in HTMLBars since we created a default global renderer
to handle components/views without a container/owner.

This change ensures that in the template only component case we fall
back to using the default component looked up from the owner. This
ensures that injections are present.

(cherry picked from commit 831c8cbbc82fa5b5b01b3eaf050b45778bf803e6)