Ensure NodeDOMTreeConstruction can be instantiated without a document.
When using Ember's `visit` API, it is possible to request that no
rendering be performed (via the `shouldRender` option). When this
happens in the browser the `DOMTreeConstruction` object still gets a
`document` passed in (because it is availble via the ambient global),
however when running in node the visit options would not include a "fake
document" (aka `SimpleDOM` document) and therefore the instantiation of
`new NodeDOMTreeConstruction(null)` would fail.

This change adds an additional method to the `DOMTreeConstruction`
class that can be overridden to make usage of the provided `document`
lazy. This does not change the behavior when using
`DOMTreeConstruction` (the `this.uselessElement` is still setup in the
constructor), but it enables the subclass that we use in Node-land to
make things work properly.