Fix usage of bound outlet name.
Previously, we were eagerly evaluating the bound reference and passing
the final value around instead of the reference to the final outlet name
value. This means that given `{{outlet foo}}` when the property `foo`
changes then no updates would be made.

The fix is to avoid eagerly evaluating the reference, band and ensure
that the `ComponentDefition` includes a tag the represents both the
outlet name and the outlet state.

While debugging this, I also came across an untested scenario where we
coerced any falsy values for outlet name into `'main'`. However, we
should only be defaulting to `'main'` if no positional argument is present.