Pass caller's self to component manager's create.
In Ember we are only using `dynamicScope` (since that is all that is
exposed), but the current `{{this}}` (aka `self`)  is also needed to
allow  us to properly target actions to their calling context.

The `dynamicScope` tracks the direct component heirarchy
without regard to blocks/yielding etc. This is a perfect analogue to the
way Ember tracks `.parentView` through a component structure.

The caller self tracks the current calling context. This means it is
aware of blocks yielding (which changes the `self` in the yielded
block), and is the correct analogue to the way Ember tracks the target
of any actions within a template block.

For example, given:

```hbs
{{#component-a}}
  {{component-b bar="bar"}}
{{/component-a}}
```

From within `component-b` the `.parentView` (derived from
`dynamicScope`) should match `component-a`, but the `targetObject`
of `component-b` should not be `component-a` (since `component-a` is not
calling it).