Refactor compileSpec into precompile since no one seems to know what a `Spec` is including me.
This teases apart meta from the root block in the
SerializedTemplate so we can apply lazy template
parsing only to the block and not the meta.

Cleanup unused tests and compile options

Add server side id support with fallback to client
side ids.

Remove destructing stuff on template render method

RenderOptions are not optional.

Add templateFactory function and tests for
templateFactory

Allow it to inject per singleton meta that is env
specific so we can use for owner injection in
Ember.

Add id to precompiled templates and node test for
default id on precompile.

Simplify `tsconfig.json`

Ensured it is working as expected by using
`--traceResolution` and `--listFiles` flags.