[BUGFIX beta] Check state of router during transition
Currently, during transitions, the state of the router is not taken into
consideration. This change makes sure that both `transitionTo()` and
`replaceWith()` examine the state of the router and uses that information
to choose the appropriate method.

If there has never been a transition since the app loaded, `replaceWith()`
will be used. If there's been a transition already `transitionTo()` will
be used instead. Regardless of which method the  developer used.