Fix {{this.foo}} !== {{foo}}
The handlebars parser simply drops the `this` for all path expressions,
but this is incorrect in Glimmer since we have the concept of local
variables (block params), so `{{this.foo}}` is semantically different
from `{{foo}}` (the former is always a “self lookup”, whereas the latter
might be a local variable lookup).

This fundamentally fixes the problem by adding `this` back into the path
expressions and allow the runtime to triage it. This also removes the
need for the “self-get” expression we previously added to address this
issue.