Ensure revalidation is not triggered for destroying components.
Fixes regression related to `willDestroyElement` causing changes to
subscribed streams. When that happened, the system would trigger
`scheduleRevalidate` on the view (even though it was being destroyed).
By the time that the async rerender occurs the queue of things to
cleanup in `_destroyingSubtreeForView` has been reset to `null` causing
the error reported.

This fix simply avoids queueing up rerenders when the component is being destroyed.