Prevent issues during destruction when `parentView` is observed.
Setting `parentView` to `null` is technically better, but unfortunately
we have not had this reset in place for a while and folks have dependent
keys that include `parentView`. Setting without `Ember.set` triggers the
mandatory setter assertion.

This removes the clearing of `parentView` (for now) so that we can
properly message it being unobservable in the future.