[DOC] Make public two-week policy for closing issues
When triaging a team, we use the _Needs Submitter Response_ to mark
issues that cannot be solved without some extra information from the
original submitter. If the submitter does not provide any extra
information after seven days, the team gently reminds the original
submitter. If another week passes, the issue might be closed.