Fix `lastIterated` not getting populated correctly
When nothing is inserted or deleted, we don't call `didInitializeChildren`
which causes `lastIterated` to not be populated. This causes us to
reiterate the list unnecessarily.

This is reported via https://github.com/emberjs/ember.js/issues/14332
and tested in https://github.com/emberjs/ember.js/pull/14347.

Unfortunately, since this relies on objects having useful validation
tags, we don't really have the infrastructure to test it in Glimmer yet.