[BUGFIX beta] Disallow accessing components’ element in FastBoot
We do not allow apps to accessing `this.element` in non-interactive
environments, such as FastBoot. This was enforced by not calling hooks
that would expose the “element” (which is often not a real DOM element)
such as `didInsertElement`. However, since we sometimes run “update”
hooks (such as `didUpdateAttrs`) that happen after the element has been
set, there is a chance that we would leak the “element” inadvertently.

This fixes the issue by making `this.element` a getter that throws in
non-interactive mode. To address cases where apps are relying on the
current accidental behavior, we added `ViewUtils.getViewElement` as a
private (“intimate”) API to forcefully retrieve the “element”. This
should balance the need for warning against incidental usage and still
provide a workaround for cases where the fake element is needed, such as
custom serializations.

As always, using private (“intimate”) APIs are strongly discouraged, not
fully supported and might not be unreliable. However, we will try our
best to keep it around and deprecate between LTS releases before
removing it.

Note: the main issue here is that the “element” you get back is likely
not a real DOM element with an undocumented set of APIs. This core issue
is not really addressed here and the supported APIs and internals of the
returned object is subject to change.

(cherry picked from commit 516baf73b6a1f4a4c4f9a8b654a2983b35c4f819)