[BUGFIX beta] Fix alias test for interior sets to consume.
This test was added to ensure that the dirtying / tagging system used by
Glimmer 2 is properly aware of an alias being dirtied. At the time it
was authored, the assumption was that we would automatically watch all
aliases but this has proven to be a performance issue when lots of
alias's are used but not consumed in the template layer.

This change updates the test to ensure that it consumes the alias before
verifying that it is dirtied.  This much more properly emulates "reality".