Do not slice `classNames` and `classNameBindings` per instance.
The vast majority of the time these arrays are completely static on the
prototype, there are relatively few instances where we actually need to
slice them but we were speculatively doing this "just in case" you
wanted it.

This change does not remove the ability to have custom `classNames` /
`classNameBindings` per instance, but it does make it so that folks
wanting to do this would need to do `this.classNames.slice()` first.

This avoids many extraneous array allocations, and just one more layer
of wasted work during initial render.