[BUGFIX beta] [FIXES #14447] fix didDestroyElement
Components via appendTo, must behave the same when torn down via destroy, and the regular rendering system tearing them down. This code-path isn’t as first class as it likely should be, as it only exists for supporting the more legacy style of unit testing. Using the component integration tests, the typical code path is taken.