Don't call getHandler from _queryParamsFor
`Router#_queryParamsFor` now takes a `handlerInfos` array instead of a route name so that it doesn't need to call `getHandler` on the Router.js instance. This is important because an asynchronously loaded Route will return a `Promise` from `getHandler`.

To support the above, several functions now take either a `TransitionState` object or an array of `HandlerInfo` objects instead of a route name and contexts/models to calculate the state.