[BUGFIX beta] Don't rely on non-cached reference
This commit fixes a bug where `OutletComponentReference`
inadvertently relied on broken behavior: it returned `null`, assuming
that that would cause the reference to be rebuilt. But
tildeio/glimmer#337 fixed the underlying bug, causing the reference to
be properly cached. As a result, we need to remember that we rendered a
`null` component, so that swapping back to the original component
replaces the empty component with the rendered component.
(cherry picked from commit 8cc93b64564ea0e7184dbab8db11b7c577973ac8)