Improve {{partial}} implementation
This commit:

Switched partials to use the new opcode building style.

Moved the evaluation of the test opcode, and the production of the
condition register, outside of the block that could be re-entered. This
means that when the expression passed to a {{partial expr}} becomes
invalidated, we no longer generate a brand new reference for the
to-boolean operation on the operand (which produces the condition).