Improve encapsulation of captured VM state
Previously, the code that captured and restored VM state manually
captured a few registers and manually restored them.

This commit puts the Frame in charge of capturing and restoring, which
means that if we want to capture more registers in the future, it will
be easy to remember to update the restoring code at the same time.