[BUGFIX lts-2-8] Fix JSHint issue with referencing `Symbol`.
The guard for using `Symbol` should not have been setup this way (`let
Symbol = Symbol || null` will never be defined AFAICT since the
transpilation would use `var` which hoists).

Either way, there is no need for the local `Symbol` variable since we
can just check if it is defined as the guard.