[BUGFIX beta] Avoid clobbering `require`.
The ember-template-compiler will use the native `crypto` library to
generate a template `id` on the server (instead of a client generated
`id` at runtime). If `require` is present and a function, it will
attempt to `require('crypto')` to generate a SHA1 for the template's
`id`.

Unfortunately, when `ember-template-compiler.js` is built it is wrapped
inside an IIFE with this `var require` in the outermost scope, which
prevents the template compiler from being able to build a template `id`
while compiling the template.

This change removes the unused `require` variable from the outermost
scope, and allows the server side template `id` generation to function
as designed.