[FEATURE beta] Assert expand properties are balanced and not nested
In #13693, it is asked to error for an unsupported feature. Though we
don't say that this is not supported, it would be nice if we could let
the user know that what he is trying to do will not work instead of
silently fail.

Fixes #13693

(cherry picked from commit 6f2d615d815fc55b5e2aa2db4a00b538ec3bc41b)