Remove mostly-unused but large build protocol
Early on, we gave every syntax element a "build" method that could be
used to generate syntax more dynamically. This added a lot of code but
never ended up being used much (the wire format is nicer to target
anyway).

This commit removes the implementations of the build protocol and
removes the one usage of it in the TestEnvironments.