Trying out a "remove syntax objects" approach
The approach in this commit is relatively naive, but the intent is to
see whether I can avoid intermediate syntax objects entirely and compile
directly from sexp -> opcodes.

I introduce a few client-side sexps that aren't in the wire format to
match the current strategy. I'm not sure if that will be necessary if
this strategy pans out.

Currently, 1,139 of 1,445 assertions in runtime pass, and I haven't
implemented block or component syntax yet, which accounts for most of
the remaining failures. The basic functionality of handlebars and
Glimmer updating, including HTML syntax, is working beautifully.