All tests pass
There are a number of important changes to how things work, and this
will not just drop in to Ember.

There is also a fair bit of cleanup left to to (type checking and small
restructuring).

Specialization and scanning is now significantly more lazy (the wire
format is kept around until a block is compiled).

Important! The wire format changed in two important ways:

1. Blocks are now nested directly where they are used, to eliminate the
   need for processing every statement ahead of time
2. Things that look like a component (dasherized) are once again
   components in the wire format, but this is much less costly now that
   blocks are always embedded locally. Since we keep around the wire
   format for longer, the fallback case is just a little re-splicing of
   the wire format, nothing too complicated.