Opcodes are now an array
This commit changes a few things ;)

First, all values referenced by opcodes directly are stored in an
environment-wide Constants pool. The opcode builder still takes strings,
expressions, blocks, etc. and converts them into constants by calling
constants.string(s), constants.expression(e), etc.

Next, an opcode is, for the moment:

  { type: OpcodeID, data: [u32, u32, u32] }

The data items point to constants. So if you expect to see an opcode
like:

  { type: 'OpenElement', data: ['div'] }

You'd instead see:

  { type: 32, data: [5, 0, 0] }

The `5` in data is a pointer into the strings pool, and `0`s are always
nulls.

The `OpenElement` opcode is defined like this:

  APPEND_OPCODES.add('OpenElement', (vm, tag) => {
    vm.stack().openElement(vm.constants.getString(tag));
  });

The `tag` passed into the evaluator is the `5` we saw before, and the evaluator
uses vm.constants.getString(tag) to get back `div` from the 5.

This structure should make it easier to optimize the opcode loop itself.

Next steps:

- use a single allocated buffer for all opcode sequences (so a block is always
  just a pair of offsets into the executable opcodes). This may be result in
  garbage if there is too much deopted code, but we can deal with that by
  intelligently allocating likely-deopted code and reusing space if it becomes
  a problem.
- use a typed array rather than a real array.
- intern strings (and possibly other value types) so we don't have many copies
  of the same string in the string table.
- use an enum for common HTML tags in the OpenElement opcode.
- other representation optimizations.
- bring back debugging info as a separate library (rather than storing things
  like debug names in the core opcodes).