Fix deopt for the field load of `.concat`
This passed all meaningful tests, but there is one test
that is asserting something I'm shocked about, that I'm going
to submit this for review anyway.  If this were true we'd do things
like call concat on a string with an array.

Which makes little sense to me.

I believe this test isn't testing a real thing.