Make `SourceLocation` mandatory in the AST types
A lot of the current code (mainly error-reporting) doesn't handle
missing loc information at all, so when you add type annotations to
their parameters you could get a type error because of this.

The choices are either to 1) start adding guards to all of those places,
or to 2) fix the cases where we didn't provide the loc information.

@wycats believe the (2) is where we want to get to, so doing (1) would
just be adding junk that has to be removed later. Therefore we decided
to omit (lie?) the current broken state from the type definition and
declare that an usage error on the production side (when we get around
to typing those spots we will be forced to deal with it instead).