[WIP] pre-allocate symbols on the pre-compiler
In this commit, all symbols are pre-allocated in the pre-compiler (aka
"the server", which is a strange name we all used to refer to the
pre-compiler, even though there is typically no actual server involved).

This means that @foo, named blocks (in usages such as `{{yield to=...}}`
and `(has-block ...)` and local variables (block params) get their
dedicated scope slots assigned and compiled into the wire-format.

TODO:

- serialize symbol table or debug map in the WF
- hook up runtime code to use the pre-built symbol table
- remove remaining runtime code that was previously doing this job
- ...eval/partials