Decouple controller lookup from URL generation.
Currently, calling `emberRouter.routerMicrolib.generate('somethign');` will
force both `something` route and `something` controller to be looked up.
We do this so that we can properly handle any query params within that
route/controller pair.

The `ember-routing-router-service` work is adding a `urlFor` method, that
will ultimately allow users to generate urls without forcing the controllers
to be looked up eagerly. Unfortunately, at the moment the eager controller
lookup is completely entangled with the `getHandler` function that we provide
to the router microlib.

This commit removes the forced eager evaluation of `_qp` on a route within the
`getHandler` function, while still supporting the current implementation of
`{{link-to}}` (which still utilizes the eager `_qp` evaluation to generate
`href`'s).