Register x-handlebars scripts into the container
Ember templates detected in the DOM have long been added to
Ember.TEMPLATES. This made them available to applications via the
Ember.DefaultResolver.

With the aim of eventually dropping the DefaultResolver, we need to add
these templates elsewhere. This commit changes them to register directly
into an application's registry. As a result, one could use a custom
resolver (without Ember.TEMPLATES support) and still leverage templates
stored in the DOM.