Update testem to allow CI to silence logging.
Travis builds are currently failing due to exceeding their limit of
4MB of log output.

testem added a new option `tap_quiet_logs` just recently (< 8 days ago).
This updates us to use the current HEAD commit of testem, so that we can
take advantage of the new feature.

It also moves the two testem configs to be `.js` files (instead of `.json`)
so that we can add logic to them if needed.