Small refactor of tracking pending requests in acceptance tests
There is no need of putting requests in an array, we're only interested
in the amount of them.

The function that checks if there is pending requests now returns a 
boolean instead of relying on zero being falsey. It was renamed to
`checkPendingRequests` to clarify that it is boolean (it’s private API)