[BUGFIX beta] Make toString safe for arrays too.
`ember-utils/to-string` is documented as a `toString` function that supports
objects without a `toString` method.

Arrays do have a string method that is based on their `join` method. This
method spec ([Array.prototype.join in exma262 2017
draft](https://tc39.github.io/ecma262/#sec-array.prototype.join)) is based on
calling `toString` on the objects that form the collection. If an object were
not to have such method, `Array.prototype.toString` would throw and error.

This PR recreates `Array.prototype.join` as closer to the spec as I could,
changing the `ToString` part with the safe version of `toString`.

Fixes #14738 and #15113.

(cherry picked from commit 2acf4186fad0285894fd0e5dfb8b68945ea32403)