[BUGFIX beta] Fix double container destroy
This code on the Engine Instance is not needed as it includes the
`ContainerProxyMixin`, which already takes care of destroying the
container.

This duplicated destroy step was causing ember-cp-validations to
attempt to destroy validators twice. This destroy process tries to
nullify some property with `o.set(‘model’, null)` on an object that is
already destroyed.
(cherry picked from commit 7e4ace6e64e6fde67be312fa1ae89b63e1d4559d)