[BUGFIX LTS] Warn about invalid properties
There was a bug in the brace expansion implementation before 2.13 that
admitted expressions like `a.{b,c` and `a.b,c` in property expansion
without failing.

On February 2017, #13231 was merged improving the implementation of
`expandProperties` and fixing the previous issue.

Sadly, people upgrading from 2.12 are having problems with this. We have
two options here:

- Backporting #13231 fixes it and can be done, as of today, without any
  problem just by cherry-picking its two commits. Sadly, this would
  break people apps in a patch version, and this bug does not seem like
  a critical bugfix.

- Adding a warning if a user is using invalid properties. I went with
  this option since it looks like the less intrusive one while letting
  the users know that their apps will break in the next update.