Ensure positional arguments are accounted for in builder.
Prior to this commit, positional args are pushed onto the stack
during `builder.component.dynamic` but we not properly `pop`'ed.

This resulted in issues when using positional param based API's
(e.g. `{{render` or `{{mount` in Ember), because the stack ordering
was off.

Tests exist for this in Ember itself (both `{{render` and `{{mount`
tests that were failing, are passing now with this change), but adding
tests here requires significant fleshing out of the testing infra.