Update DOMOperations to allow creation of the useless element to be customized.
`NodeDOMTreeConstruction` inherits from `DOMOperations` and intentionally overrides
`setupUselessElement` (to be a noop). This is required in Ember (currently) because
we create an instance of `NodeDOMTreeConstruction` even when we don't have a `document`
(e.g. `App.visit('/', { shouldRender: false {})`). That is obviously not ideal (
`document` should never be `null` here), and that issue will be fixed upstream.

This commit can be reverted once `shouldRender: false` no longer instantiates the
Glimmer environment.